learn process only for learn
